fix: rename MDX files to follow kebab-case convention
- Rename zh_Cn_PrivacyPlolicyContent.mdx → zh-cn-privacy-policy-content.mdx
- Rename en_PrivacyPolicyContent.mdx → en-privacy-policy-content.mdx
- Rename zh_CN_agreement.mdx → zh-cn-agreement.mdx
- Rename en_agreement.mdx → en-agreement.mdx
- Update import statements in page.tsx files to match new file names
