+++ src/cook-web/src/app/agreement/Contents/en-agreement.mdx
