+++ src/cook-web/src/app/agreement/Contents/zh-cn-agreement.mdx
