+++ src/cook-web/src/app/privacy/contents/en-privacy-policy-content.mdx
